refactor(domain): 패키지 이동
패키지 이동을 하였습니다.
